Change static routing example names + doc
