+++ ipmininet/examples/README.md
@@ -20,6 +20,10 @@ The following sections will detail the topologies.
    - [StaticAddressNetwork](#staticaddressnetwork)
    - [PartialStaticAddressNet](#partialstaticaddressnetwork)
    - [StaticRoutingNet](#staticroutingnet)
+   - [StaticRoutingNetBasic](#staticroutingnetbasic)
+   - [StaticRoutingNetIntermediate](#staticroutingnetintermediate)
+   - [StaticRoutingNetComplex](#staticroutingnetcomplex)
+   - [StaticRoutingNetFailure](#staticroutingnetfailure)
    - [SpanningTreeNet](#spanningtreenet)
    - [SpanningTreeHub](#spanningtreehub)
    - [SpanningTreeBus](#spanningtreebus)
@@ -191,6 +195,34 @@ _args_ : n/a
 
 This network uses static routes with zebra and static
 daemons.
+
+## StaticRoutingNetBasic
+
+_topo name_ : static_routing_network_basic
+_args_ : n/a
+
+This nework uses static routes with zebra and static daemons.
+This topology uses only 4 routers.
+
+## StaticRoutingNetIntermediate
+
+_topo name_ : static_routing_network_intermediate
+_args_ : n/a
+
+This network uses static routes with zebra and static daemons.
+This topology uses 6 routers. The routes are not the same as
+if they were chosen by OSPF6, but a path from X to Y is the
+exact opposite to the path from Y to X.
+
+## StaticRoutingNetComplex
+
+_topo name_ : static_routing_network_complex
+_args_ : n/a
+
+This network uses static routes with zebra and static daemons.
+This topology uses 6 routers. The routes are not the same
+as if they were chosen by OSPF6. The path from X to Y
+and its reverse path are not the same.
 
 ## StaticRoutingNetFailure
