+++ ipmininet/examples/__main__.py
@@ -15,11 +15,11 @@ from .sshd import SSHTopo
 from .router_adv_network import RouterAdvNet
 from .simple_openr_network import SimpleOpenrNet
 from .static_address_network import StaticAddressNet
-from .static_routing_network_3 import StaticRoutingNet3
-from .static_routing_network_4 import StaticRoutingNet4
+from .static_routing_network_intermediate import StaticRoutingNetIntermediate
+from .static_routing_network_complex import StaticRoutingNetComplex
 from .partial_static_address_network import PartialStaticAddressNet
 from .static_routing import StaticRoutingNet
-from .static_routing_network_2 import StaticRoutingNet2
+from .static_routing_network_basic import StaticRoutingNetBasic
 from .static_routing_failure import StaticRoutingNetFailure
 from .spanning_tree import SpanningTreeNet
 from .spanning_tree_intermediate import SpanningTreeIntermediate
@@ -41,11 +41,11 @@ TOPOS = {'simple_ospf_network': SimpleOSPFNet,
          'router_adv_network': RouterAdvNet,
          'simple_openr_network': SimpleOpenrNet,
          'static_address_network': StaticAddressNet,
-         'static_routing_network_3': StaticRoutingNet3,
-         'static_routing_network_4': StaticRoutingNet4,
+         'static_routing_network_intermediate': StaticRoutingNetIntermediate,
+         'static_routing_network_complex': StaticRoutingNetComplex,
          'partial_static_address_network': PartialStaticAddressNet,
          'static_routing_network': StaticRoutingNet,
-         'static_routing_network_2': StaticRoutingNet2,
+         'static_routing_network_basic': StaticRoutingNetBasic,
          'static_routing_network_failure': StaticRoutingNetFailure,
          'spanning_tree_network': SpanningTreeNet,
          'spanning_tree_intermediate': SpanningTreeIntermediate,
